+++ data_source/pixiv/__init__.py
@@ -1,0 +1,3 @@
+from .pixiv_data_source import PixivDataSource
+
+__all__ = ("PixivDataSource",)
