+++ data_source/pixiv/cache_manager.py
@@ -4,17 +4,9 @@ import typing
 
 
 class CacheManager:
-    _semaphore: asyncio.Semaphore
-    _waiting: typing.Dict[typing.Any, asyncio.Future]
-
-    def __init__(self, simultaneous_query: int = 4,
-                 loop: typing.Optional[asyncio.AbstractEventLoop] = None):
-        if sys.version_info >= (3, 10, 0):
-            self._semaphore = asyncio.Semaphore(
-                value=simultaneous_query)  # 限制从远程获取的并发量为8
-        else:
-            self._semaphore = asyncio.Semaphore(
-                value=simultaneous_query, loop=loop)
+    def __init__(self, simultaneous_query: int = 4):
+        self._semaphore = asyncio.Semaphore(
+            value=simultaneous_query)  # 用于限制从远程获取的并发量
         self._waiting = {}
 
     T = typing.TypeVar("T")
@@ -22,7 +14,7 @@ class CacheManager:
     async def get(self, identifier: typing.Any,
                   cache_loader: typing.Callable[[], typing.Coroutine[typing.Any, typing.Any, T]],
                   remote_fetcher: typing.Callable[[], typing.Coroutine[typing.Any, typing.Any, T]],
-                  cache_updater: typing.Callable[[], typing.Coroutine],
+                  cache_updater: typing.Callable[[T], typing.Coroutine[typing.Any, typing.Any, typing.NoReturn]],
                   timeout: typing.Optional[int] = 0) -> T:
         cache = await cache_loader()
         if cache is not None:
@@ -31,9 +23,9 @@ class CacheManager:
         if identifier in self._waiting:
             return await self._waiting[identifier]
 
+        fut = asyncio.Future()
+        self._waiting[identifier] = fut
         try:
-            fut = asyncio.Future()
-            self._waiting[identifier] = fut
             asyncio.create_task(self._fetch(
                 fut, remote_fetcher, cache_updater, timeout))
             return await fut
@@ -42,7 +34,7 @@ class CacheManager:
 
     async def _fetch(self, fut: asyncio.Future,
                      remote_fetcher: typing.Callable[[], typing.Coroutine[typing.Any, typing.Any, T]],
-                     cache_updater: typing.Callable[[T], typing.Coroutine],
+                     cache_updater: typing.Callable[[T], typing.Coroutine[typing.Any, typing.Any, typing.NoReturn]],
                      timeout: typing.Optional[float] = None):
         await self._semaphore.acquire()
         try:
