+++ data_source/pixiv/pkg_context.py
@@ -1,4 +1,4 @@
-from ..pkg_context import Context
 from ..pkg_context import context as parent_context
+from ..pkg_context import Context
 
 context = Context(parent=parent_context)
